Refactor FantasyTeam.Standings
* Extract SeasonEnded module
* Function to add season ended not related to standings
* Organize private functions
* Closes #270
